Refinements to pipeline

+++ database/clickhouse/schema/3101_participant_perk_values.sql
@@ -5,25 +5,25 @@ CREATE TABLE IF NOT EXISTS game_data.participant_perk_values
     teamid UInt8,
     puuid FixedString (78),
 
-    primary_var1_1 UInt16,
-    primary_var2_1 UInt16,
-    primary_var3_1 UInt16,
-    primary_var1_2 UInt16,
-    primary_var2_2 UInt16,
-    primary_var3_2 UInt16,
-    primary_var1_3 UInt16,
-    primary_var2_3 UInt16,
-    primary_var3_3 UInt16,
-    primary_var1_4 UInt16,
-    primary_var2_4 UInt16,
-    primary_var3_4 UInt16,
+    primary_var1_1 Int32,
+    primary_var2_1 Int32,
+    primary_var3_1 Int32,
+    primary_var1_2 Int32,
+    primary_var2_2 Int32,
+    primary_var3_2 Int32,
+    primary_var1_3 Int32,
+    primary_var2_3 Int32,
+    primary_var3_3 Int32,
+    primary_var1_4 Int32,
+    primary_var2_4 Int32,
+    primary_var3_4 Int32,
 
-    sub_var1_1 UInt16,
-    sub_var2_1 UInt16,
-    sub_var3_1 UInt16,
-    sub_var1_2 UInt16,
-    sub_var2_2 UInt16,
-    sub_var3_2 UInt16
+    sub_var1_1 Int32,
+    sub_var2_1 Int32,
+    sub_var3_1 Int32,
+    sub_var1_2 Int32,
+    sub_var2_2 Int32,
+    sub_var3_2 Int32
 )
 ENGINE = MergeTree
 ORDER BY (matchid, teamid, puuid, run_id);
